Add files via upload
This draft is a selection of my notes on Rep(G) as a unitary fusion category and a graphical technique to compute its F symbols, accumulated over my research assistantship with the Center for Topological Quantum Computing at New York University Abu Dhabi. This project was inspired because of Steve Simon's book Topological Quantum (2022). In particular, Steve comments in at the end of Chapter 20 in his book that there is a lack of thorough written work on Rep(G) aimed at physicists, and little documentation on how to compute 6j symbols of finite groups.